fixup! DTLS 1.3 SSL Listener Demo

+++ demos/dtlsecho/main.c
@@ -312,7 +312,7 @@ int main(int argc, char **argv)
         server_addr = BIO_ADDRINFO_address(res);
 
         /* Connect the UDP socket to the server (sets default peer address) */
-        if (!BIO_connect(client_skt, server_addr, 0)) {
+        if (!BIO_connect((int)client_skt, server_addr, 0)) {
             perror("Unable to UDP connect to server");
             BIO_ADDRINFO_free(res);
             goto exit;
